forgotten cast

+++ lstchain/scripts/lstchain_r0g_to_r0v.py
@@ -162,7 +162,7 @@ def main():
                 new_status = np.where(ordered_pix_mask,
                                       pixel_status | PixelStatus.DVR_STATUS_0,
                                       pixel_status & UNSET_DVR_BIT_MASK)
-                event.pixel_status.data = np.uint8(new_status).tobytes()
+                event.pixel_status.data = new_status.tobytes()
 
                 stream.write_message(event)
